Rename icons and schemas
